Move MockEvent to its own file
so it's easier to re-use it from other tests

Since this class referenced Agenda::DATE_IN_THE_FUTURE, I also
needed to move it to another file, and moved all dates declaration
for consistency
